+++ tools/python-uv/tests/versions_test.rs
@@ -29,4 +29,42 @@ mod python_uv_tool {
         assert!(output.aliases.contains_key("latest"));
         assert_eq!(output.aliases.get("latest"), output.latest.as_ref());
     }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn parses_uv_toml() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_plugin("uv-test").await;
+
+        assert_eq!(
+            plugin
+                .parse_version_file(ParseVersionFileInput {
+                    content: r#"required-version = "==2.1""#.into(),
+                    file: "uv.toml".into(),
+                    ..Default::default()
+                })
+                .await,
+            ParseVersionFileOutput {
+                version: Some(UnresolvedVersionSpec::parse("=2.1").unwrap()),
+            }
+        );
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn parses_pyproject_toml() {
+        let sandbox = create_empty_proto_sandbox();
+        let plugin = sandbox.create_plugin("uv-test").await;
+
+        assert_eq!(
+            plugin
+                .parse_version_file(ParseVersionFileInput {
+                    content: "[tool.uv]\nrequired-version = \"~=2.1\"".into(),
+                    file: "pyproject.toml".into(),
+                    ..Default::default()
+                })
+                .await,
+            ParseVersionFileOutput {
+                version: Some(UnresolvedVersionSpec::parse("~2.1").unwrap()),
+            }
+        );
+    }
 }